updated todo list
fixed include paths to the best of my knowledge.

i believe im rendering a depth buffer thats blank. color output is red, and im unsure of why it is this way.

may be to do with shadow depth maps, and im hoping to fix it tonight. took a short break but im back.